Decompile sky0f1228d0

+++ src/include/game/sky.h
@@ -17,7 +17,7 @@ f32 skyRound(f32 value);
 void skyChooseCloudVtxColour(struct skything18 *arg0, f32 arg1);
 void sky0f11f6ec(struct skything18 *arg0, f32 arg1);
 Gfx *skyRender(Gfx *gdl);
-void sky0f1228d0(struct skything18 *arg0, Mtxf *arg1, s32 arg2, f32 arg3, f32 arg4, struct skything38 *arg5);
+void sky0f1228d0(struct skything18 *arg0, Mtxf *arg1, u16 arg2, f32 arg3, f32 arg4, struct skything38 *arg5);
 bool sky0f122ce8(struct skything38 *arg0, struct skything38 *arg1);
 Gfx *sky0f122d4c(Gfx *gdl, struct skything38 *arg1, struct skything38 *arg2, struct skything38 *arg3, f32 arg4, bool arg5);
 Gfx *sky0f123fd4(Gfx *gdl, struct skything38 *arg1, struct skything38 *arg2, struct skything38 *arg3, struct skything38 *arg4, f32 arg5);
